minimum amount of address restriction

+++ frontend/src/app/features/mainLayout/account/addressSettings/controller.js
@@ -2,6 +2,7 @@ import find from 'lodash/find';
 import merge from 'lodash/merge';
 import remove from 'lodash/remove';
 import { componentName as addressForm } from './components/addressForm/component';
+import { componentName as genericModal } from 'components/genericModal/component';
 
 /* @ngInject */
 export default class AccountAddresses {
@@ -92,18 +93,30 @@ export default class AccountAddresses {
     deleteAddress(address) {
         const id = address.id;
 
-        this.addressService.deleteEntity({ id })
-            .then(resp => {
-                if (resp.data.success) {
-                    this.notificationsService.success({ msg: 'Address Deleted' });
-                    remove(this.customer.addresses, item => {
-                        return item.id === id;
-                    });
-                } else {
-                    this.notificationsService.alert({ msg: resp.data.message });
+        if (this.customer.addresses.length === 1) {
+            const modalInstance = this.$uibModal.open({
+                animation: true,
+                component: genericModal,
+                resolve: {
+                    title: () => 'Cannot Delete Address',
+                    text: () => 'You must have at least one address on file',
+                    confirmButtonLabel: () => 'OK'
                 }
-            }, err => {
-                this.notificationsService.alert({ msg: err.message });
             });
+        } else {
+            this.addressService.deleteEntity({ id })
+                .then(resp => {
+                    if (resp.data.success) {
+                        this.notificationsService.success({ msg: 'Address Deleted' });
+                        remove(this.customer.addresses, item => {
+                            return item.id === id;
+                        });
+                    } else {
+                        this.notificationsService.alert({ msg: resp.data.message });
+                    }
+                }, err => {
+                    this.notificationsService.alert({ msg: err.message });
+                });
+        }
     }
 }
